requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pinecone-client==3.0.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
langchain==0.1.5
langchain-openai==0.0.5
langchain-pinecone==0.0.3
tiktoken==0.5.2
